# openapi.v2.legacy.yaml
openapi: 3.0.0
info:
  title: AirTrust API
  description: Sistema de Gestão de Treinamentos e Qualificações Aeronáuticas
  version: 2.0.0
  contact:
    name: AirTrust Support
    email: suporte@airtrust.com.br
servers:
  - url: http://localhost:8787
    description: Desenvolvimento
  - url: https://api.airtrust.com.br
    description: Produção
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Funcionario:
      type: object
      required:
        - nome
        - matricula
      properties:
        id:
          type: integer
          readOnly: true
        nome:
          type: string
          minLength: 3
        matricula:
          type: string
        cpf:
          type: string
          pattern: '^\d{11}$'
        email:
          type: string
          format: email
        telefone:
          type: string
        funcao:
          type: string
        base:
          type: string
        status:
          type: string
          enum: [ATIVO, INATIVO, AFASTADO]
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
security:
  - bearerAuth: []
paths:
...
  /api/v2/funcionarios:
    get:
      summary: Listar Funcionários
      tags: [Funcionários]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de funcionários
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  funcionarios:
                    type: array
                    items:
                      $ref: '#/components/schemas/Funcionario'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Criar Funcionário
      tags: [Funcionários]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Funcionario'
      responses:
        '201':
          description: Funcionário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Funcionario'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...